fix some escaped errors
